+++ CHANGELOG.md
@@ -6,7 +6,11 @@ master
   Next, execute `rake ember:heroku`. [#544]
 * Generate an empty `yarn.lock` so that Heroku understands that the
   application's deployment target requires `yarn`. Closes [#538]. [#540]
+* No longer support `rails < 4.2`. [#543]
+* Generate an empty `yarn.lock` so that Heroku understands that the
+  application's deployment target requires `yarn`. Closes [#538]. [#540]
 
+[#543]: https://github.com/thoughtbot/ember-cli-rails/pull/543
 [#544]: https://github.com/thoughtbot/ember-cli-rails/pull/544
 [#538]: https://github.com/thoughtbot/ember-cli-rails/issues/538
 [#540]: https://github.com/thoughtbot/ember-cli-rails/pull/540